Avoid empty `instancePath` when `propertyNames` is violated.
This makes `_validate` take an `Instance` object instead of dynamic.
And instead forces the callers to convert any values to `Instance`
objects.

This is probably a good idea because:
 * Caller is the only one who can construct the `instancePath`.
 * It avoids a lot of unnecessary `instance is! Instance` checks.

Fixing this it was discovered that `propertyNames` errors were reported
with an empty `instancePath`. Which could cause confusion for anyone
trying to display the error.